+++ x/evm/embeds/hardhat.config.js
@@ -2,5 +2,10 @@ require("@nomicfoundation/hardhat-toolbox");
 
 /** @type import('hardhat/config').HardhatUserConfig */
 module.exports = {
-  solidity: "0.8.24",
+  solidity: {
+    compilers: [
+      { version: "0.4.19" },
+      { version: "0.8.24" },
+    ],
+  },
 };